# [WIP] New opt-in partitioning strategy that may help with read optimization…