--- a/accumulo/src/main/scala/geotrellis/spark/io/accumulo/AccumuloLayerReindexer.scala
+++ b/accumulo/src/main/scala/geotrellis/spark/io/accumulo/AccumuloLayerReindexer.scala
@@ -75,9 +75,9 @@ class AccumuloLayerReindexer(
     val layerCopier = AccumuloLayerCopier(attributeStore, layerReader, layerWriter)
 
     layerWriter.write(tmpId, layerReader.read[K, V, M](id), keyIndex)
-    layerDeleter.delete(id)
+    layerDeleter.delete[K](id)
     layerCopier.copy[K, V, M](tmpId, id)
-    layerDeleter.delete(tmpId)
+    layerDeleter.delete[K](tmpId)
   }
 
   def reindex[
@@ -99,8 +99,8 @@ class AccumuloLayerReindexer(
     val layerCopier = AccumuloLayerCopier(attributeStore, layerReader, layerWriter)
 
     layerWriter.write(tmpId, layerReader.read[K, V, M](id), keyIndexMethod.createIndex(existingKeyIndex.keyBounds))
-    layerDeleter.delete(id)
+    layerDeleter.delete[K](id)
     layerCopier.copy[K, V, M](tmpId, id)
-    layerDeleter.delete(tmpId)
+    layerDeleter.delete[K](tmpId)
   }
 }