--- a/cassandra/src/main/resources/reference.conf
+++ b/cassandra/src/main/resources/reference.conf
@@ -13,19 +13,41 @@
 # limitations under the License.
 
 geotrellis.cassandra {
-  port                 = 9042
-  catalog              = "metadata"
-  keyspace             = "geotrellis"
-  replicationStrategy  = "SimpleStrategy"
-  replicationFactor    = 1
-  localDc              = "datacenter1"
+  port = 9042
+  catalog = "metadata"
+  keyspace = "geotrellis"
+  replicationStrategy = "SimpleStrategy"
+  replicationFactor = 1
+  localDc = "datacenter1"
   usedHostsPerRemoteDc = 0
   allowRemoteDCsForLocalConsistencyLevel = false
   threads {
     collection.read = default
     rdd {
       write = default
-      read  = default
+      read = default
     }
   }
 }
+
+# Comment out the block above and uncomment this to run the Cassandra tests
+# using the new 'read-optimized-partitioner' strategy. 
+#geotrellis.cassandra {
+#  port                 = 9042
+#  catalog              = "metadata"
+#  keyspace             = "geotrellis"
+#  replicationStrategy  = "SimpleStrategy"
+#  replicationFactor    = 1
+#  localDc              = "datacenter1"
+#  usedHostsPerRemoteDc = 0
+#  allowRemoteDCsForLocalConsistencyLevel = false
+#  partitionStrategy    = "read-optimized-partitioner"
+#  tilesPerPartition    = 64
+#  threads {
+#    collection.read = default
+#    rdd {
+#      write = default
+#      read  = default
+#    }
+#  }
+#}