--- a/cassandra/src/main/resources/reference.conf
+++ b/cassandra/src/main/resources/reference.conf
@@ -31,7 +31,7 @@ geotrellis.cassandra {
 }
 
 # Comment out the block above and uncomment this to run the Cassandra tests
-# using the new 'read-optimized-partitioner' strategy. 
+# using the new 'read-optimized-partitioner' strategy.
 #geotrellis.cassandra {
 #  port                 = 9042
 #  catalog              = "metadata"