--- a/cassandra/src/main/resources/reference.conf
+++ b/cassandra/src/main/resources/reference.conf
@@ -21,6 +21,15 @@ geotrellis.cassandra {
   localDc = "datacenter1"
   usedHostsPerRemoteDc = 0
   allowRemoteDCsForLocalConsistencyLevel = false
+  # 'writeoptimized' or 'readoptimized', default is 'writeoptimized'.
+  # For 'readoptimized', clients can configure the maximum size of a partition
+  # read based on tiles in the partition.  Careful tuning of this value can
+  # lead to optimized read performance when using key/partition caching.
+#  indexStrategy = "writeoptimized"
+  # Only relevant if partitionStrategy is 'readoptimized' - configures the
+  # maximum number of tiles to store per parition.
+  indexStrategy = "readoptimized"
+  tilesPerParition = 64
   threads {
     collection.read = default
     rdd {
@@ -29,25 +38,3 @@ geotrellis.cassandra {
     }
   }
 }
-
-# Comment out the block above and uncomment this to run the Cassandra tests
-# using the new 'read-optimized-partitioner' strategy.
-#geotrellis.cassandra {
-#  port                 = 9042
-#  catalog              = "metadata"
-#  keyspace             = "geotrellis"
-#  replicationStrategy  = "SimpleStrategy"
-#  replicationFactor    = 1
-#  localDc              = "datacenter1"
-#  usedHostsPerRemoteDc = 0
-#  allowRemoteDCsForLocalConsistencyLevel = false
-#  partitionStrategy    = "read-optimized-partitioner"
-#  tilesPerPartition    = 64
-#  threads {
-#    collection.read = default
-#    rdd {
-#      write = default
-#      read  = default
-#    }
-#  }
-#}