--- a/cassandra/src/main/resources/reference.conf
+++ b/cassandra/src/main/resources/reference.conf
@@ -25,11 +25,10 @@ geotrellis.cassandra {
   # For 'readoptimized', clients can configure the maximum size of a partition
   # read based on tiles in the partition.  Careful tuning of this value can
   # lead to optimized read performance when using key/partition caching.
-#  indexStrategy = "writeoptimized"
+  indexStrategy = "writeoptimized"
   # Only relevant if partitionStrategy is 'readoptimized' - configures the
   # maximum number of tiles to store per parition.
-  indexStrategy = "readoptimized"
-  tilesPerParition = 64
+#  tilesPerParition = 64
   threads {
     collection.read = default
     rdd {