--- a/cassandra/src/main/scala/geotrellis/spark/io/cassandra/CassandraLayerManager.scala
+++ b/cassandra/src/main/scala/geotrellis/spark/io/cassandra/CassandraLayerManager.scala
@@ -29,7 +29,7 @@ import scala.reflect.ClassTag
 
 class CassandraLayerManager(attributeStore: CassandraAttributeStore, instance: CassandraInstance)(implicit sc: SparkContext)
     extends LayerManager[LayerId]{
-  def delete[K: AvroRecordCodec: Boundable: JsonFormat: ClassTag](id: LayerId): Unit =
+  def delete[K: ClassTag](id: LayerId): Unit =
     CassandraLayerDeleter(attributeStore, instance).delete[K](id)
 
   def copy[