… cassandra/src/main/scala/geotrellis/spark/io/cassandra/CassandraLayerReindexer.scala
… cassandra/src/main/scala/geotrellis/spark/io/cassandra/CassandraLayerWriter.scala
… cassandra/src/main/scala/geotrellis/spark/io/cassandra/CassandraRDDReader.scala
… cassandra/src/main/scala/geotrellis/spark/io/cassandra/CassandraRDDWriter.scala
… cassandra/src/main/scala/geotrellis/spark/io/cassandra/CassandraValueReader.scala
… cassandra/src/main/scala/geotrellis/spark/io/cassandra/conf/CassandraConfig.scala
… cassandra/src/test/scala/geotrellis/spark/CassandraTestEnvironment.scala
… cassandra/src/test/scala/geotrellis/spark/io/cassandra/bench/CassandraIndexStrategySpec.scala
… docs/cassandra/CassandraIndexStrategySpec.scala
… docs/cassandra/cassandra-optimization.md
… docs/cassandra/cassandra-optimization.rst
… docs/cassandra/cassandra-test.md
… docs/cassandra/cassandra-test.rst
… docs/guide/tile-backends.rst
… hbase/src/main/scala/geotrellis/spark/io/hbase/HBaseLayerDeleter.scala
… hbase/src/main/scala/geotrellis/spark/io/hbase/HBaseLayerManager.scala
… hbase/src/main/scala/geotrellis/spark/io/hbase/HBaseLayerReindexer.scala
… s3/src/main/scala/geotrellis/spark/io/s3/S3LayerDeleter.scala
… spark-etl/src/main/scala/geotrellis/spark/etl/config/json/ConfigFormats.scala
… spark-testkit/src/main/scala/geotrellis/spark/testkit/io/PersistenceSpec.scala
… spark/src/main/scala/geotrellis/spark/io/GenericLayerMover.scala
… spark/src/main/scala/geotrellis/spark/io/GenericLayerReindexer.scala
… spark/src/main/scala/geotrellis/spark/io/LayerDeleter.scala
… spark/src/main/scala/geotrellis/spark/io/LayerManager.scala
… spark/src/main/scala/geotrellis/spark/io/file/FileLayerDeleter.scala
… spark/src/main/scala/geotrellis/spark/io/file/FileLayerManager.scala
… spark/src/main/scala/geotrellis/spark/io/hadoop/HadoopLayerDeleter.scala
… spark/src/main/scala/geotrellis/spark/io/hadoop/HadoopLayerManager.scala
… spark/src/main/scala/geotrellis/spark/io/index/KeyIndex.scala