… for cassandra by taking better advantage of the SFC locality guarantees.

cc @pomadchin 

## Overview

Strawman PR for comments/feedback... code is still quite rough but looking for thoughts on indexing changes and API changes required.

Addresses thoughts in #2831 

### Checklist

- [ ] `docs/CHANGELOG.rst` updated, if necessary  (not sure if necessary?)
- [x] `docs` guides update, if necessary
- [ ] New user API has useful Scaladoc strings (existing layer API doesn't seem to)
- [ ] Unit tests added for bug-fix or new feature (TBD - need to exercise both index strategies in cassandra test)

### Notes

I'm hoping that:
  1. You can help me avoid having to grab the range of the key via the `KeyIndex`... Is there another way to access this information that I'm missing?  If so then I think we can avoid the change to the `LayerManager` and `LayerDeleter` APIs.  My suspicion is that I've missed something fundamental here about the abstraction, but I'm not quite sure what. 

  2. You can help me understand the type signature of `KeyIndex.indexRanges` as I'm not sure I'm making proper usage of it.  I'm guessing that, due to the fact that an SFC may have "breaks" in it where there are large contiguous blocks of its range that are not used, that the `Seq[(BigInt, BigInt)]` return type is representing something like this: 

  |(begin SFC index range)---------|(unused part of range index)*******|(more used index)--------|(more unused index)******|------(end SFC index range)|

  so that really what the `(BigInt, BigInt)` tuples are capturing is the coverage area over the range of the SFC like this: 

  |------|*****|------|****|------|
  ....0..............1............2...

  `Seq[(0th tuple, 1st tuple, 2nd tuple)]`

  is this correct?

  3. You can point me in the right direction for doing a real performance analysis to compare the differences in the partition strategies.  I tried running the geotrellis/cassandra test cases both with and without the read-optimized-partitioner I introduced here and it doesn't seem to break anything... but I also noticed most of the tests only seem to operate on Zoom Level 1 (based on observations from cqlsh) and they're all mixed workloads (read/write), so it's hard to say if there's much benefit one way or the other.  I did add some timing code into the test spec itself as a _very_ rough way of benchmarking and it looks like at worst we've not done much harm: 

With the following (updated) configuration (note the new "read-optimized-partitioner"): 

```
geotrellis.cassandra {
  port                 = 9042
  catalog              = "metadata"
  keyspace             = "geotrellis"
  replicationStrategy  = "SimpleStrategy"
  replicationFactor    = 1
  localDc              = "datacenter1"
  usedHostsPerRemoteDc = 0
  allowRemoteDCsForLocalConsistencyLevel = false
  partitionStrategy    = "read-optimized-partitioner"
  tilesPerPartition    = 64
  threads {
    collection.read = default
    rdd {
      write = default
      read  = default
    }
  }
}
```

I got these timings from the test spec: 

```
18/12/21 08:00:40 INFO CassandraTileFeatureSpaceTimeSpec: Execution Time: 100655ms
18/12/21 08:01:16 INFO CassandraSpatialSpec: Execution Time: 31812ms
18/12/21 08:01:16 INFO CassandraLayerProviderSpec: Execution Time: 538ms
18/12/21 08:01:47 INFO CassandraTileFeatureSpatialSpec: Execution Time: 30784ms
18/12/21 08:03:20 INFO CassandraSpaceTimeSpec: Execution Time: 93133ms
18/12/21 08:03:21 INFO CassandraAttributeStoreSpec: Execution Time: 457ms
```

And with this (original) configuration:

```
geotrellis.cassandra {
  port = 9042
  catalog = "metadata"
  keyspace = "geotrellis"
  replicationStrategy = "SimpleStrategy"
  replicationFactor = 1
  localDc = "datacenter1"
  usedHostsPerRemoteDc = 0
  allowRemoteDCsForLocalConsistencyLevel = false
  threads {
    collection.read = default
    rdd {
      write = default
      read = default
    }
  }
}
```

I got these timings from the test spec: 

```
18/12/21 08:09:23 INFO CassandraTileFeatureSpaceTimeSpec: Execution Time: 95980ms
18/12/21 08:09:58 INFO CassandraSpatialSpec: Execution Time: 31297ms
18/12/21 08:09:59 INFO CassandraLayerProviderSpec: Execution Time: 1135ms
18/12/21 08:10:29 INFO CassandraTileFeatureSpatialSpec: Execution Time: 29870ms
18/12/21 08:11:59 INFO CassandraSpaceTimeSpec: Execution Time: 89329ms
18/12/21 08:11:59 INFO CassandraAttributeStoreSpec: Execution Time: 411ms
```

All of the differences nominally seem to me to be within the margin of error... if anything I've made things a little slower, which only makes sense... more data = more time to write it :-) 

Anyway - I'd be interested to know if you have any thoughts on how to more properly benchmark this.

Addresses #2831