@ALPSMAC 2018-12-21
Sorry about the PR churn... was fighting with gpg to get the signatures to work correctly.