@ALPSMAC 2019-01-02
@pomadchin FYI - I'm back from the holidays and am looking at code cleanup now.  I'll also take a look at how to do a better performance benchmark with results that I can believe a little bit more to convince ourselves we didn't introduce a feature that makes things worse ;-)