@pomadchin 2019-02-20
@ALPSMAC have you tried to run some tests on a real Cassandra cluster?